requirements.txt
telethon>=1.24.0
colorama>=0.4.4
python-dotenv>=0.19.2
